django_API_REST_BASE/site_002/prerequirement.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
djangorestframework==3.9.0
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
Pygments==2.3.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
uritemplate==3.0.0
urllib3==1.24.1
